check_compliance.py: Fix License and Codeowners checks
The 'sh' library seems to have some magic for iterating over lines,
which the License check relied on (maybe accidentally). The check broke
when the 'sh' library was removed in commit 3f6c5f9
("check_compliance.py/what_changed.py: Remove 'sh' library dependency
and fix pylint warnings"):

    Traceback (most recent call last):
      File ".../check_compliance.py", line 1210, in main
        n_fails = _main(args)
      File ".../check_compliance.py", line 1156, in _main
        test.run()
      File ".../check_compliance.py", line 780, in run
        copyfile(file, copy)
      File ".../shutil.py", line 114, in copyfile
        with open(src, 'rb') as fsrc:
    FileNotFoundError: [Errno 2] No such file or directory: 'b'

(The 'b' there is the first character of a filename.)

Fix it by running .splitlines() on the reusult of git().

Make a similar change to the Codeowners check, which looks like it might
have the same issue.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>